requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
pycryptodome==3.19.0
gunicorn==21.2.0
